% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.sens.R
\name{par.sens}
\alias{par.sens}
\title{Parameter sensitivity}
\usage{
par.sens(parm, par.range, iteration, steps, mate.comp, fixed_parms, all)
}
\arguments{
\item{par.range}{Range of parameter values over which to test sensitivity.}

\item{steps}{The number of steps (e.g. years) per iteration.}

\item{par}{The parameter for which to test sensitivity over one or more values.}

\item{...}{other parameters from SongEvo()

The function currently allows examination of only one parameter at a time and requires at least two iterations.}
}
\value{
array named sens.results. The sens.results array from par.sens(), which includes summary.results from SongEvo() for a range of parameter values.  summary.results from SongEvo() includes population summary values for each time step (dimension 1) in each iteration (dimension 2) of the model.  Population summary values are contained in five additional dimensions: population size for each time step of each iteration (“sample.n”), the population mean and variance of the song feature studied (“trait.pop.mean” and “trait.pop.variance”), with associated lower (“lci”) and upper (“uci”) confidence intervals.
}
\description{
This function allows testing the sensitivity of SongEvo to different parameter values.
}
