% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{song.data}
\alias{song.data}
\title{White Crown Sparrow Song Observations}
\format{A data frame with 89 rows and 3 variables:
\describe{
  \item{Population}{Locality of the observed male sparrow}
  \item{Year}{Year in which samples were taken}
  \item{Trill.FBW}{Mean observed trill bandwidth}
}}
\source{
\url{http://FIXME/}
}
\usage{
song.data
}
\description{
A dataset of mean trill bandwidth from 3 WCSP male populations in 1969 and 2005
}
\keyword{datasets}
